cache: refactor disk_cache.go to separate layers and keep interface implementations clean.
