+++ pkg/chunk/disk_cache.go
@@ -17,11 +17,8 @@
 package chunk
 
 import (
-	"encoding/binary"
 	"errors"
 	"fmt"
-	"hash/crc32"
-	"hash/fnv"
 	"io"
 	"io/fs"
 	"os"
@@ -29,7 +26,6 @@ import (
 	"reflect"
 	"regexp"
 	"runtime"
-	"sort"
 	"strconv"
 	"strings"
 	"sync"
@@ -38,13 +34,9 @@ import (
 	"time"
 
 	"github.com/charlievieth/fastwalk"
-	"github.com/davies/groupcache/consistenthash"
 	"github.com/dustin/go-humanize"
 	"github.com/google/uuid"
 	"github.com/juicedata/juicefs/pkg/utils"
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/twmb/murmur3"
-	"github.com/vmihailenco/msgpack/v5"
 )
 
 var (
@@ -52,7 +44,6 @@ var (
 	cacheDir            = "raw"
 	maxIODur            = time.Second * 30
 	stagingBlocks       atomic.Int64
-	errNotCached        = errors.New("not cached")
 	errStageFull        = errors.New("space not enough on device")
 	errStageConcurrency = errors.New("concurrent staging limit reached")
 )
@@ -1065,519 +1056,4 @@ func (cache *cacheStore) scanStaging() {
 	if count > 0 {
 		logger.Infof("Found %d staging blocks (%s) in %s with %s", count, humanize.IBytes(usage), cache.dir, time.Since(start))
 	}
-}
-
-type cacheManager struct {
-	sync.Mutex
-	consistentMap *consistenthash.Map
-	storeMap      map[string]*cacheStore
-	stores        []*cacheStore
-	metrics       *cacheManagerMetrics
-}
-
-func legacyKeyHash(s string) uint32 {
-	hash := fnv.New32()
-	_, _ = hash.Write([]byte(s))
-	return hash.Sum32()
-}
-
-// hasMeta reports whether path contains any of the magic characters
-// recognized by Match.
-func hasMeta(path string) bool {
-	magicChars := `*?[`
-	if runtime.GOOS != "windows" {
-		magicChars = `*?[\`
-	}
-	return strings.ContainsAny(path, magicChars)
-}
-
-var osPathSeparator = string([]byte{os.PathSeparator})
-
-func expandDir(pattern string) []string {
-	pattern = strings.TrimRight(pattern, "/")
-	if runtime.GOOS == "windows" {
-		pattern = strings.TrimRight(pattern, osPathSeparator)
-	}
-	if pattern == "" {
-		return []string{"/"}
-	}
-	if !hasMeta(pattern) {
-		return []string{pattern}
-	}
-	dir, f := filepath.Split(pattern)
-	if hasMeta(f) {
-		matched, err := filepath.Glob(pattern)
-		if err != nil {
-			logger.Errorf("glob %s: %s", pattern, err)
-			return []string{pattern}
-		}
-		return matched
-	}
-	var rs []string
-	for _, p := range expandDir(dir) {
-		rs = append(rs, filepath.Join(p, f))
-	}
-	return rs
-}
-
-type CacheManager interface {
-	cache(key string, p *Page, force, dropCache bool)
-	remove(key string, staging bool)
-	load(key string) (ReadCloser, error)
-	exist(key string) (string, bool)
-	uploaded(key string, size int)
-	stage(key string, data []byte, tierID uint8) (string, error)
-	removeStage(key string) error
-	stats() (int64, int64)
-	usedMemory() int64
-	isEmpty() bool
-	getMetrics() *cacheManagerMetrics
-}
-
-func newCacheManager(config *Config, reg prometheus.Registerer, uploader func(key, path string, force bool) bool) CacheManager {
-	getEnvs()
-	metrics := newCacheManagerMetrics(reg)
-	if config.CacheDir == "memory" || !config.CacheEnabled() {
-		return newMemStore(config, metrics)
-	}
-	var dirs []string
-	for _, d := range utils.SplitDir(config.CacheDir) {
-		dd := expandDir(d)
-		if config.AutoCreate {
-			dirs = append(dirs, dd...)
-		} else {
-			for _, d := range dd {
-				if fi, err := os.Stat(d); err == nil && fi.IsDir() {
-					dirs = append(dirs, d)
-				}
-			}
-		}
-	}
-	if len(dirs) == 0 {
-		config.CacheSize = 100 << 20
-		logger.Warnf("No cache dir existed, use memory cache instead, cache size: 100 MiB")
-		return newMemStore(config, metrics)
-	}
-	sort.Strings(dirs)
-	dirCacheSize := int64(config.CacheSize) / int64(len(dirs))
-	dirCacheItems := config.CacheItems / int64(len(dirs))
-	m := &cacheManager{
-		consistentMap: consistenthash.New(100, murmur3.Sum32),
-		storeMap:      make(map[string]*cacheStore, len(dirs)),
-		stores:        make([]*cacheStore, len(dirs)),
-		metrics:       metrics,
-	}
-
-	// 20% of buffer could be used for pending pages
-	pendingPages := int(config.BufferSize) * 2 / 10 / config.BlockSize / len(dirs)
-	for i, d := range dirs {
-		store := newCacheStore(metrics, strings.TrimSpace(d)+string(filepath.Separator), dirCacheSize, dirCacheItems, pendingPages, config, uploader)
-		m.stores[i] = store
-		m.storeMap[store.id] = store
-		m.consistentMap.Add(store.id)
-	}
-	go m.cleanup()
-	return m
-}
-
-func (m *cacheManager) getMetrics() *cacheManagerMetrics {
-	return m.metrics
-}
-
-func (m *cacheManager) cleanup() {
-	for !m.isEmpty() {
-		var ids []string
-		m.Lock()
-		for id, s := range m.storeMap {
-			if s == nil || !s.available() {
-				ids = append(ids, id)
-			}
-		}
-		m.Unlock()
-		for _, id := range ids {
-			m.removeStore(id)
-		}
-		time.Sleep(time.Second)
-	}
-}
-
-func (m *cacheManager) isEmpty() bool {
-	return m.length() == 0
-}
-
-func (m *cacheManager) length() int {
-	m.Lock()
-	defer m.Unlock()
-	return len(m.storeMap)
-}
-
-func (m *cacheManager) removeStore(id string) {
-	m.Lock()
-	m.consistentMap.Remove(id)
-	var dir string
-	if s := m.storeMap[id]; s != nil {
-		dir = s.dir
-	}
-	delete(m.storeMap, id)
-	for i, c := range m.stores {
-		if c != nil && c.id == id {
-			m.stores[i] = nil
-		}
-	}
-	m.Unlock()
-	logger.Errorf("cache dir `%s`(%s) is unavailable, removed", dir, id)
-}
-
-func (m *cacheManager) getStore(key string) *cacheStore {
-	for {
-		m.Lock()
-		id := m.consistentMap.Get(key)
-		s := m.storeMap[id]
-		m.Unlock()
-		if s == nil || s.available() {
-			return s
-		}
-		m.removeStore(id)
-	}
-}
-
-func (m *cacheManager) removeStage(key string) error {
-	if s := m.getStore(key); s == nil {
-		return errCacheDown
-	} else {
-		return s.removeStage(key)
-	}
-}
-
-// Deprecated: use getStore instead
-func (m *cacheManager) getStoreLegacy(key string) *cacheStore {
-	return m.stores[legacyKeyHash(key)%uint32(len(m.stores))]
-}
-
-func (m *cacheManager) usedMemory() int64 {
-	var used int64
-	for _, s := range m.stores {
-		if s != nil {
-			used += s.usedMemory()
-		}
-	}
-	return used
-}
-
-func (m *cacheManager) stats() (int64, int64) {
-	var cnt, used int64
-	for _, s := range m.stores {
-		if s != nil {
-			c, u := s.stats()
-			cnt += c
-			used += u
-		}
-	}
-	return cnt, used
-}
-
-func (m *cacheManager) cache(key string, p *Page, force, dropCache bool) {
-	store := m.getStore(key)
-	if store != nil {
-		store.cache(key, p, force, dropCache)
-	}
-}
-
-type ReadCloser interface {
-	// io.Reader
-	io.ReaderAt
-	io.Closer
-}
-
-func (m *cacheManager) load(key string) (ReadCloser, error) {
-	store := m.getStore(key)
-	if store == nil {
-		return nil, errors.New("no available cache dir")
-	}
-	r, err := store.load(key)
-	if err == errNotCached {
-		legacy := m.getStoreLegacy(key)
-		if legacy != store && legacy != nil {
-			r, err = legacy.load(key)
-		}
-	}
-	return r, err
-}
-
-func (m *cacheManager) exist(key string) (string, bool) {
-	store := m.getStore(key)
-	if store == nil {
-		return "", false
-	}
-	loc := store.dir
-	existed, err := store.exist(key)
-	if err == errNotCached {
-		legacy := m.getStoreLegacy(key)
-		if legacy != store && legacy != nil {
-			existed, _ = legacy.exist(key)
-			loc = legacy.dir
-		}
-	}
-	return loc, existed
-}
-
-func (m *cacheManager) remove(key string, staging bool) {
-	store := m.getStore(key)
-	if store != nil {
-		store.remove(key, staging)
-	}
-}
-
-func (m *cacheManager) stage(key string, data []byte, tierID uint8) (string, error) {
-	store := m.getStore(key)
-	if store != nil {
-		return store.stage(key, data, tierID)
-	}
-	return "", errors.New("no available cache dir")
-}
-
-func (m *cacheManager) uploaded(key string, size int) {
-	store := m.getStore(key)
-	if store != nil {
-		store.uploaded(key, size)
-	}
-}
-
-/* --- Checksum --- */
-const (
-	CsNone   = "none"
-	CsFull   = "full"
-	CsShrink = "shrink"
-	CsExtend = "extend"
-
-	csBlock = 32 << 10
-)
-
-var crc32c = crc32.MakeTable(crc32.Castagnoli)
-
-const maxTierID = 3 // maximum valid tier id
-
-// stageFooter is msgpack metadata appended at the end of a stage file:
-//
-//   - file content: [data] [checksum?] [footer]
-//   - footer: [magic uint16] [len uint16] [msgpack data]
-//
-// We use trailer length parity to indicate
-// checksum presence (checksum bytes are always multiple of 4):
-//   - with checksum: msgpack length is padded to a multiple of 4
-//   - without checksum: msgpack length is padded to a non-multiple of 4
-//
-// Therefore, trailer length (checksum + msgpack) is multiple of 4 iff checksum
-// exists, so openCacheFile can distinguish the two and find the trailing
-// msgpack metadata.
-type stageFooter struct {
-	Tier uint8  `msgpack:"tier"`
-	Pad  []byte `msgpack:"pad"` // pad to make msgpack length parity match checksum presence
-}
-
-func (f *stageFooter) marshal(align bool) ([]byte, error) {
-	var p int
-	switch r := stageFooterBaseLen % 4; {
-	case align:
-		p = (4 - r) % 4 // pad up to the next multiple of 4
-	case r == 0:
-		p = 1 // break the multiple-of-4 alignment
-	}
-	f.Pad = stageFooterPad[p]
-	data, err := msgpack.Marshal(f)
-	if err != nil {
-		return nil, err
-	}
-
-	buff := make([]byte, 0, 4+len(data))
-	buff = binary.BigEndian.AppendUint16(buff, stageFooterMagic)
-	buff = binary.BigEndian.AppendUint16(buff, uint16(len(data)))
-	buff = append(buff, data...)
-	return buff, nil
-}
-
-func (f *stageFooter) unmarshal(cf *cacheFile) error {
-	if cf.footerOff == 0 {
-		f.Tier = 0 // no footer, fall back to the default
-		return nil
-	}
-	var hdr [4]byte
-	if _, err := cf.File.ReadAt(hdr[:], cf.footerOff); err != nil {
-		return fmt.Errorf("read stage footer header: %w", err)
-	}
-	if magic := binary.BigEndian.Uint16(hdr[:2]); magic != stageFooterMagic {
-		return fmt.Errorf("invalid stage footer magic %#04x", magic)
-	}
-	size := int64(binary.BigEndian.Uint16(hdr[2:4]))
-	if size == 0 {
-		return fmt.Errorf("invalid stage footer length %d", size)
-	}
-	if cf.footerOff+4+size != cf.size {
-		return fmt.Errorf("stage footer size mismatch: offset %d, size %d, file size %d", cf.footerOff, size, cf.size)
-	}
-	buf := make([]byte, size)
-	if _, err := cf.File.ReadAt(buf, cf.footerOff+4); err != nil {
-		return fmt.Errorf("read stage footer data: %w", err)
-	}
-	if err := msgpack.Unmarshal(buf, f); err != nil {
-		return err
-	}
-	if f.Tier > maxTierID {
-		f.Tier = 0 // ignore unknown/corrupted tier, fall back to the default
-	}
-	return nil
-}
-
-var (
-	stageFooterBaseLen int
-	stageFooterPad     [4][]byte
-	stageFooterMagic   = uint16(0x4653)
-)
-
-func init() {
-	b, err := msgpack.Marshal(&stageFooter{Pad: make([]byte, 0)})
-	if err != nil {
-		panic(err)
-	}
-	stageFooterBaseLen = len(b)
-
-	for i := 0; i < len(stageFooterPad); i++ {
-		stageFooterPad[i] = make([]byte, i)
-	}
-}
-
-type cacheFile struct {
-	*os.File
-	length    int // length of data
-	csLevel   string
-	size      int64 // size of file
-	footerOff int64 // offset of stageFooter in file (0 if none)
-}
-
-// Calculate 32-bits checksum for every 32 KiB data, so 512 Bytes for 4 MiB in total
-func checksum(data []byte) []byte {
-	length := len(data)
-	buf := utils.NewBuffer(uint32((length-1)/csBlock+1) * 4)
-	for start, end := 0, 0; start < length; start = end {
-		end = start + csBlock
-		if end > length {
-			end = length
-		}
-		sum := crc32.Checksum(data[start:end], crc32c)
-		buf.Put32(sum)
-	}
-	return buf.Bytes()
-}
-
-func openCacheFile(name string, length int, level string) (*cacheFile, error) {
-	fp, err := os.Open(name)
-	if err != nil {
-		return nil, err
-	}
-	fi, err := fp.Stat()
-	if err != nil {
-		_ = fp.Close()
-		return nil, err
-	}
-	checksumLength := int64(((length-1)/csBlock + 1) * 4)
-	extra := fi.Size() - int64(length)
-	if extra < 0 {
-		_ = fp.Close()
-		return nil, fmt.Errorf("invalid file size %d, data length %d", fi.Size(), length)
-	}
-	footerOff := int64(0)
-	switch {
-	case extra == 0:
-		level = CsNone // data only, no footer
-	case extra%4 == 0:
-		switch {
-		case extra == checksumLength:
-		case extra > checksumLength:
-			footerOff = int64(length) + checksumLength
-		default:
-			_ = fp.Close()
-			return nil, fmt.Errorf("invalid file size %d, data length %d, checksum length %d", fi.Size(), length, checksumLength)
-		}
-	default:
-		level = CsNone
-		footerOff = int64(length)
-	}
-	cf := &cacheFile{File: fp, length: length, csLevel: level, footerOff: footerOff, size: fi.Size()}
-	return cf, nil
-}
-
-func (cf *cacheFile) ReadAt(b []byte, off int64) (n int, err error) {
-	logger.Tracef("CacheFile length %d level %s, readat off %d buffer size %d", cf.length, cf.csLevel, off, len(b))
-	defer func() {
-		logger.Tracef("CacheFile readat returns n %d err %s", n, err)
-	}()
-	if cf.csLevel == CsNone || cf.csLevel == CsFull && (off != 0 || len(b) != cf.length) {
-		return cf.File.ReadAt(b, off)
-	}
-	var rb = b     // read buffer
-	var roff = off // read offset
-	if cf.csLevel == CsExtend {
-		roff = off / csBlock * csBlock
-		rend := int(off) + len(b)
-		if rend%csBlock != 0 {
-			rend = (rend/csBlock + 1) * csBlock
-			if rend > cf.length {
-				rend = cf.length
-			}
-		}
-		if size := rend - int(roff); size != len(b) {
-			p := NewOffPage(size)
-			rb = p.Data
-			defer func() {
-				if err == nil {
-					n = copy(b, rb[off-roff:])
-				} else {
-					n = 0
-				}
-				p.Release()
-			}()
-		}
-	}
-	if n, err = cf.File.ReadAt(rb, roff); err != nil {
-		return
-	}
-
-	ioff := int(roff) / csBlock // index offset
-	if cf.csLevel == CsShrink {
-		if roff%csBlock != 0 {
-			if o := csBlock - int(roff)%csBlock; len(rb) <= o {
-				return
-			} else {
-				rb = rb[o:]
-				ioff += 1
-			}
-		}
-		if end := int(roff) + n; end != cf.length && end%csBlock != 0 {
-			if len(rb) <= end%csBlock {
-				return
-			}
-			rb = rb[:len(rb)-end%csBlock]
-		}
-	}
-	// now rb contains the data to check
-	length := len(rb)
-	buf := utils.NewBuffer(uint32((length-1)/csBlock+1) * 4)
-	if _, err = cf.File.ReadAt(buf.Bytes(), int64(cf.length+ioff*4)); err != nil {
-		logger.Warnf("Read checksum of data length %d checksum offset %d: %s", length, cf.length+ioff*4, err)
-		return
-	}
-	for start, end := 0, 0; start < length; start = end {
-		end = start + csBlock
-		if end > length {
-			end = length
-		}
-		sum := crc32.Checksum(rb[start:end], crc32c)
-		expect := buf.Get32()
-		logger.Debugf("Cache file read data start %d end %d checksum %d, expected %d", start, end, sum, expect)
-		if sum != expect {
-			err = fmt.Errorf("data checksum %d != expect %d", sum, expect)
-			break
-		}
-	}
-	return
 }
